Implement radio button and organize renderer modules
