+++ src/ui/widget.rs
@@ -4,12 +4,14 @@ mod row;
 pub mod button;
 pub mod checkbox;
 pub mod panel;
+pub mod radio;
 pub mod text;
 
 pub use button::Button;
 pub use checkbox::Checkbox;
 pub use column::Column;
 pub use panel::Panel;
+pub use radio::Radio;
 pub use row::Row;
 pub use text::Text;
