+++ src/ui/widget/radio.rs
@@ -7,28 +7,29 @@ use crate::ui::{
     Row, Text, Widget,
 };
 
-pub struct Checkbox<M, R> {
-    is_checked: bool,
-    on_toggle: Box<Fn(bool) -> M>,
+pub struct Radio<M, R> {
+    is_selected: bool,
+    on_click: M,
     label: String,
     renderer: std::marker::PhantomData<R>,
 }
 
-impl<M, R> Checkbox<M, R> {
-    pub fn new<F>(is_checked: bool, label: &str, f: F) -> Self
+impl<M, R> Radio<M, R> {
+    pub fn new<F, V>(value: V, label: &str, selected: Option<V>, f: F) -> Self
     where
-        F: Fn(bool) -> M + 'static,
+        V: Eq + Copy,
+        F: Fn(V) -> M + 'static,
     {
-        Checkbox {
-            is_checked,
-            on_toggle: Box::new(f),
+        Radio {
+            is_selected: Some(value) == selected,
+            on_click: f(value),
             label: String::from(label),
             renderer: std::marker::PhantomData,
         }
     }
 }
 
-impl<'a, M, R> Widget<'a> for Checkbox<M, R>
+impl<'a, M, R> Widget<'a> for Radio<M, R>
 where
     R: Renderer + column::Renderer + text::Renderer + 'static,
     M: Copy,
@@ -61,7 +62,7 @@ where
                 state: ButtonState::Pressed,
             } => {
                 if mouse_over {
-                    messages.push((self.on_toggle)(!self.is_checked));
+                    messages.push(self.on_click);
                 }
             }
             _ => {}
@@ -87,7 +88,7 @@ where
         );
 
         (renderer as &mut Renderer).draw(
-            self.is_checked,
+            self.is_selected,
             children[0].bounds(),
             text_bounds,
             cursor_position,
@@ -102,19 +103,19 @@ where
 pub trait Renderer {
     fn draw(
         &mut self,
-        is_checked: bool,
+        is_selected: bool,
         bounds: Rectangle<f32>,
         label_bounds: Rectangle<f32>,
         cursor_position: Point,
     ) -> MouseCursor;
 }
 
-impl<'a, M, R> From<Checkbox<M, R>> for Element<'a, M, R>
+impl<'a, M, R> From<Radio<M, R>> for Element<'a, M, R>
 where
     R: Renderer + column::Renderer + text::Renderer + 'static,
     M: Copy + 'static,
 {
-    fn from(checkbox: Checkbox<M, R>) -> Element<'a, M, R> {
+    fn from(checkbox: Radio<M, R>) -> Element<'a, M, R> {
         Element::new(checkbox)
     }
 }
